Rename modules in BioMASS models (#115)
* Rename files

* Fix file description

* Fix note
