+++ biomass/models/Nakakuki_Cell_2010/observable.py
@@ -5,7 +5,7 @@ import numpy as np
 from biomass.dynamics.solver import get_steady_state, solve_ode
 
 from .name2idx import C, V
-from .set_model import DifferentialEquation
+from .ode import DifferentialEquation
 
 
 class Observable(DifferentialEquation):
