+++ biomass/models/insulin_signaling/search_param.py
@@ -3,7 +3,7 @@ import numpy as np
 from biomass.estimation import convert_scale, initialize_search_param
 
 from .name2idx import C, V
-from .set_model import initial_values, param_values
+from .ode import initial_values, param_values
 
 
 class SearchParam(object):
