+++ biomass/models/nfkb_pathway/__init__.py
@@ -1,5 +1,5 @@
-from .fitness import OptimizationProblem
 from .name2idx import C, V
+from .ode import initial_values, param_values
+from .problem import OptimizationProblem
 from .reaction_network import ReactionNetwork
-from .set_model import initial_values, param_values
 from .viz import Visualization
