+++ biomass/models/tgfb_smad/observable.py
@@ -3,7 +3,7 @@ import numpy as np
 from biomass.dynamics.solver import solve_ode
 
 from .name2idx import C, V
-from .set_model import DifferentialEquation
+from .ode import DifferentialEquation
 
 
 class Observable(DifferentialEquation):
